+++ xivo_sysconf/modules/network.py
@@ -56,31 +56,6 @@ def rename_ethernet_interface(args):
         NETLOCK.release()
 
 
-SWAP_ETH_SCHEMA = xys.load("""
-name1: !~~prefixedDec eth
-name2: !~~prefixedDec eth
-""")
-
-
-def swap_ethernet_interfaces(args):
-    """
-    POST /swap_ethernet_interfaces
-
-    args ex:
-    {'name1': "eth0",
-     'name2': "eth1"}
-    """
-    if not xys.validate(args, SWAP_ETH_SCHEMA):
-        raise HttpReqError(415, "invalid arguments for command")
-    if not NETLOCK.acquire_write(NET_LOCK_TIMEOUT):
-        raise HttpReqError(503, "unable to take NETLOCK for writing after %s seconds" % NET_LOCK_TIMEOUT)
-    try:
-        xivo_config.swap_ethernet_interfaces(args['name1'], args['name2'])
-        return True
-    finally:
-        NETLOCK.release()
-
-
 def _val_modify_network_config(args):
     """
     ad hoc validation function for modify_network_config command
@@ -119,4 +94,3 @@ def modify_network_config(args):
 
 http_json_server.register(network_config, CMD_R)
 http_json_server.register(rename_ethernet_interface, CMD_RW)
-http_json_server.register(swap_ethernet_interfaces, CMD_RW)